# Optimize variant price and is published product filters