I want to merge this change because it optimizes the filtering of products by "is published" and "variant price" filters.

is_published before ([450ms](https://explain.dalibo.com/plan/8a5e22cb7gd7efe7))
is_published after ([1,65ms](https://explain.dalibo.com/plan/5b2ee83779h3b5a5)) - 272x faster
variant_price before ([230ms](https://explain.dalibo.com/plan/ch8c11356e7ee3bc))
variant_price after ([3,68ms](https://explain.dalibo.com/plan/bbbcf92ed03488eb)) - 62,5x faster

<!-- Please mention all relevant issue numbers. -->

# Impact

- [ ] New migrations
- [ ] New/Updated API fields or mutations
- [ ] Deprecated API fields or mutations
- [ ] Removed API types, fields, or mutations

# Docs

<!-- Docs are stored in a separate repository: https://github.com/saleor/saleor-docs/. -->
<!-- Please provide a link to the PR that updates documentation for your changes. -->
<!-- If changes in docs are not required, please mention that in the description. -->

- [ ] Link to documentation:

# Pull Request Checklist

<!-- Please keep this section. It will make the maintainer's life easier. -->

- [ ] Privileged queries and mutations are either absent or guarded by proper permission checks
- [ ] Database queries are optimized and the number of queries is constant
- [ ] Database migrations are either absent or optimized for zero downtime
- [ ] The changes are covered by test cases
- [ ] All new fields/inputs/mutations have proper labels added (`ADDED_IN_X`, `PREVIEW_FEATURE`, etc.)
- [ ] All migrations have proper dependencies
- [ ] All indexes are added concurrently in migrations
- [ ] All RunSql and RunPython migrations have revert option defined