@tomaszszymanski129 2024-10-15
@aniav there are two things I need to check here to make sure filtering behaviour is exactly as it was, though I switched to more urgent stuff and never had the time. I'll try to test it this week and if it's fine I'll merge. If there are some changes that needs to be applied it has to wait in queue. In 2nd case, I'll write an update so we don't see it's hanging open for no reason.